Separate ZGC-dependent test from JvmGcMetricsTest
